+++ lib/src/utils/souceUtils.d.ts
@@ -9,12 +9,7 @@ export declare function shrinkPermissionSets(permissionsets: any): Promise<void>
 export declare function profileElementInjection(profiles: any, ensureObjectPermissionsFromAdmin?: {
     ensureObjectPermissions: any;
 }, customObjectsFilter?: any[]): Promise<void>;
-export declare function moveSourceFilesByFolder(): Promise<Array<{
-    from: string;
-    to: string;
-}>>;
 export declare function logFixes(updatedfiles: any): Promise<void>;
-export declare function logMoves(movedSourceFiles: any): Promise<void>;
 export declare function getConnectionFromArgv(): Promise<argvConnection>;
 export declare function applyFixes(tags: any, root?: any, filter?: any[]): Promise<aggregatedFixResults>;
 export declare function applySourceFixes(filter: string[]): Promise<aggregatedFixResults>;
@@ -27,5 +22,5 @@ declare type fixResult = {
     operation: string;
     message: string;
 };
-export declare function updateProfiles(profiles: any, retrievePackageDir: any, forceSourcePull: any): Promise<void>;
+export declare function updateProfiles(profiles: any, customObjects: any, forceSourcePull: any): Promise<void>;
 export {};