+++ src/commands/jayree/automation/ltngsync/status.ts
@@ -1,15 +1,16 @@
 /*
- * Copyright (c) 2020, jayree
+ * Copyright (c) 2021, jayree
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { core, flags, SfdxCommand } from '@salesforce/command';
+import { flags, SfdxCommand } from '@salesforce/command';
+import { Messages, Connection } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import puppeteer from 'puppeteer';
 
-core.Messages.importMessagesDirectory(__dirname);
-const messages = core.Messages.loadMessages('sfdx-jayree', 'ltngsyncstatus');
+Messages.importMessagesDirectory(__dirname);
+const messages = Messages.loadMessages('sfdx-jayree', 'ltngsyncstatus');
 export default class LtngSyncStatus extends SfdxCommand {
   public static description = messages.getMessage('commandDescription');
 
@@ -189,7 +190,7 @@ export default class LtngSyncStatus extends SfdxCommand {
     return { tables, userContactsItem, userEventsItem };
   }
 
-  private async login(conn: core.Connection, page: puppeteer.Page) {
+  private async login(conn: Connection, page: puppeteer.Page) {
     await page.goto(conn.instanceUrl + '/secur/frontdoor.jsp?sid=' + conn.accessToken, {
       waitUntil: 'networkidle2',
     });