+++ src/commands/jayree/flowtestcoverage.ts
@@ -1,15 +1,16 @@
 /*
- * Copyright (c) 2020, jayree
+ * Copyright (c) 2021, jayree
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { core, SfdxCommand } from '@salesforce/command';
+import { SfdxCommand } from '@salesforce/command';
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 
-core.Messages.importMessagesDirectory(__dirname);
+Messages.importMessagesDirectory(__dirname);
 
-const messages = core.Messages.loadMessages('sfdx-jayree', 'flowtestcoverage');
+const messages = Messages.loadMessages('sfdx-jayree', 'flowtestcoverage');
 
 export default class FlowTestCoverage extends SfdxCommand {
   public static description = messages.getMessage('commandDescription');