+++ src/commands/jayree/packagedescription/create.ts
@@ -1,16 +1,17 @@
 /*
- * Copyright (c) 2020, jayree
+ * Copyright (c) 2021, jayree
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { core, flags, SfdxCommand } from '@salesforce/command';
+import { flags, SfdxCommand } from '@salesforce/command';
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import AdmZip from 'adm-zip';
-import * as convert from 'xml-js';
+import { builder } from '../../../utils/xml';
 
-core.Messages.importMessagesDirectory(__dirname);
-const messages = core.Messages.loadMessages('sfdx-jayree', 'createpackagedescription');
+Messages.importMessagesDirectory(__dirname);
+const messages = Messages.loadMessages('sfdx-jayree', 'createpackagedescription');
 
 export default class CreatePackageDescription extends SfdxCommand {
   // hotfix to receive only one help page
@@ -50,22 +51,14 @@ export default class CreatePackageDescription extends SfdxCommand {
     const text = this.flags.description.replace(/\\n/g, '\n');
 
     const fileContentjs = {
-      _declaration: { _attributes: { version: '1.0', encoding: 'utf-8' } },
-      Package: [
-        {
-          _attributes: { xmlns: 'http://soap.sforce.com/2006/04/metadata' },
-          description: text,
-          version: '43.0',
-        },
-      ],
+      Package: {
+        $: { xmlns: 'http://soap.sforce.com/2006/04/metadata' },
+        description: [text],
+        version: ['52.0'],
+      },
     };
 
-    newZip.addFile(
-      'unpackaged/package.xml',
-      Buffer.from(convert.js2xml(fileContentjs, { compact: true, spaces: 4 })),
-      '',
-      0o644
-    );
+    newZip.addFile('unpackaged/package.xml', Buffer.from(builder.buildObject(fileContentjs)), '', 0o644);
 
     newZip.writeZip(inputfile);
     // this.ux.log(newZip.getEntries()[0].header.toString());