+++ src/commands/jayree/packagedescription/remove.ts
@@ -1,16 +1,17 @@
 /*
- * Copyright (c) 2020, jayree
+ * Copyright (c) 2021, jayree
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { core, flags, SfdxCommand } from '@salesforce/command';
+import { flags, SfdxCommand } from '@salesforce/command';
+import { Messages } from '@salesforce/core';
 import { AnyJson } from '@salesforce/ts-types';
 import AdmZip from 'adm-zip';
-import * as convert from 'xml-js';
+import { builder, parseStringSync } from '../../../utils/xml';
 
-core.Messages.importMessagesDirectory(__dirname);
-const messages = core.Messages.loadMessages('sfdx-jayree', 'removepackagedescription');
+Messages.importMessagesDirectory(__dirname);
+const messages = Messages.loadMessages('sfdx-jayree', 'removepackagedescription');
 
 export default class RemovePackageDescription extends SfdxCommand {
   // hotfix to receive only one help page
@@ -53,26 +54,19 @@ export default class RemovePackageDescription extends SfdxCommand {
       if (fileName.includes('package.xml')) {
         let fileContentjs;
         const fileTXTContent = zip.readAsText(fileName);
-        const xml = convert.xml2js(fileTXTContent, { compact: true });
-        if ('description' in xml['Package']) {
-          text = xml['Package']['description']['_text'];
+        const xml = parseStringSync(fileTXTContent);
+        if (xml.Package.description && xml.Package.description.length > 0) {
+          text = xml.Package.description.toString();
           action = 'removed';
           this.ux.log(action + ' description: ' + text);
           fileContentjs = {
-            _declaration: {
-              _attributes: { version: '1.0', encoding: 'utf-8' },
+            Package: {
+              $: { xmlns: 'http://soap.sforce.com/2006/04/metadata' },
+              types: xml.Package.types,
+              version: xml.Package.version,
             },
-            Package: [
-              {
-                _attributes: {
-                  xmlns: 'http://soap.sforce.com/2006/04/metadata',
-                },
-                types: xml['Package']['types'],
-                version: xml['Package']['version'],
-              },
-            ],
           };
-          newZip.addFile(fileName, Buffer.from(convert.js2xml(fileContentjs, { compact: true, spaces: 4 })), '', 0o644);
+          newZip.addFile(fileName, Buffer.from(builder.buildObject(fileContentjs)), '', 0o644);
         } else {
           action = '';
           this.ux.log('no description found');