+++ src/hooks/prerun.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, jayree
+ * Copyright (c) 2021, jayree
  * All rights reserved.
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
@@ -42,7 +42,7 @@ export const prerun: Hook<'prerun'> = async function (options) {
         storedServerMaxRevisionCounter = serverMaxRevisionCounter;
         // eslint-disable-next-line no-empty
       } catch {}
-      if (storedServerMaxRevisionCounter) {
+      if (storedServerMaxRevisionCounter >= 0) {
         try {
           cli.log(`Reset local tracking files to revision ${storedServerMaxRevisionCounter}.`);
           await execa('sfdx', [